autonomous: round 8 improvements
- Fix _compute_raw_preflash returning 1D shape (3,) instead of (1,1,3)
  when preflash is disabled, preventing silent broadcasting bugs
- Preserve input dtype in ResizingService.small_preview to avoid
  float64 promotion of float32 pipeline data
- Clamp crop size to image dimensions in crop_image to prevent
  empty/negative crops when crop size exceeds image size

+++ src/spektrafilm/runtime/services/resize.py
@@ -39,6 +39,6 @@ class ResizingService:
                 scale_factor,
                 channel_axis=2,
                 order=0,
-            )
+            ).astype(image.dtype)
         return image
